Por suerte agarramos algunso errores
