+++ scr/download.py
@@ -1,22 +1,12 @@
 import argparse
 
-from lib.parameter import ClientParameter, OutputVerbosity, SendMethod
-
-class CustomHelpFormatter(argparse.HelpFormatter):
-    def _format_action_invocation(self, action):
-        if action.option_strings:
-            parts = []
-            for option_string in action.option_strings:
-                parts.append(option_string)
-            return ', '.join(parts)
-        else:
-            return super()._format_action_invocation(action)
+from lib.parameter import ClientParameter, OutputVerbosity, SendMethod, CustomFormatter
 
 def obtainParameters():
     parser = argparse.ArgumentParser(
         prog = "download", 
         description = "Default description",
-        formatter_class=CustomHelpFormatter,
+        formatter_class=CustomFormatter,
     )
 
     verbosity = parser.add_mutually_exclusive_group()
@@ -34,7 +24,7 @@ def obtainParameters():
     )
 
     parser.add_argument("-H", "--host", default="", dest="addr", help="server IP address")
-    parser.add_argument("-p", "--port", default="", dest="port", type=int, help="server port")
+    parser.add_argument("-p", "--port", default=123123, dest="port", type=int, help="server port")
 
     parser.add_argument("-d", "--dst", default="", dest="filepath", help="destination file path")
     parser.add_argument("-n", "--name", default="", dest="filename", help="name file name")
@@ -68,7 +58,7 @@ def obtainParameters():
         port = args.port,
         filePath = args.filepath,
         nameFile = args.filename,
-        method = method.selective_repeat if method.stop_wait is None else method.stop_wait
+        method = args.select_repeat if args.stop_wait is None else args.stop_wait
     )
 
 def main(parameter):
