+++ scr/lib/parameter.py
@@ -1,3 +1,4 @@
+import argparse
 from dataclasses import dataclass
 from enum import Enum
 
@@ -26,4 +27,14 @@ class ServerParameter:
     port: int
     storagePath: str
     method: SendMethod
+
+class CustomFormatter(argparse.HelpFormatter):
+    def _format_action_invocation(self, action):
+        if action.option_strings:
+            parts = []
+            for option_string in action.option_strings:
+                parts.append(option_string)
+            return ', '.join(parts)
+        else:
+            return super()._format_action_invocation(action)
